Extract Media class to own file
